# Benchmarks when not tracing