For a reality check to see how we are doing, it is essential
that we compare the overhead of work that's done without tracing
entirely. The results are a dunk of one's head in cold water.

1. tracing, sampling: NeverSample vs AlwaysSample
```shell
$ benchstat w_nosample.txt w_sample.txt
name  old time/op    new time/op    delta
It       200ns ± 2%     611ns ± 1%  +205.12%  (p=0.029 n=4+4)

name  old alloc/op   new alloc/op   delta
It        128B ± 0%      576B ± 0%  +350.00%  (p=0.029 n=4+4)

name  old allocs/op  new allocs/op  delta
It        2.00 ± 0%      4.00 ± 0%  +100.00%  (p=0.029 n=4+4)
```

2. no tracing vs NeverSample
```shell
$ benchstat notracing_reuse.txt w_nosample.txt
name  old time/op    new time/op    delta
It      5.68ns ± 2%  200.25ns ± 2%  +3427.08%  (p=0.029 n=4+4)

name  old alloc/op   new alloc/op   delta
It      0.00B ±NaN%   128.00B ± 0%      +Inf%  (p=0.029 n=4+4)

name  old allocs/op  new allocs/op  delta
It       0.00 ±NaN%      2.00 ± 0%      +Inf%  (p=0.029 n=4+4)
```

3. no tracing vs AlwaysSample
```shell
$ benchstat notracing_reuse.txt w_sample.txt
name  old time/op    new time/op    delta
It      5.68ns ± 2%  611.00ns ± 1%  +10661.78%  (p=0.029 n=4+4)

name  old alloc/op   new alloc/op   delta
It      0.00B ±NaN%   576.00B ± 0%       +Inf%  (p=0.029 n=4+4)

name  old allocs/op  new allocs/op  delta
It       0.00 ±NaN%      4.00 ± 0%       +Inf%  (p=0.029 n=4+4)
```