@ramonza 2018-03-16
You can make the overhead of tracing (percentage-wise) almost arbitrarily large by making "doWork" cheaper. Conversely, you can easily "fix" the high overhead of tracing by replacing "doWork" with something arbitrarily more expensive. 

The point of microbenchmarks is to gradually improve the performance of a small piece of code, not to prove that overhead is acceptable. You can't prove that in a microbenchmark because, as shown above, it all depends on what else your program is doing.