@rakyll 2018-03-16
I agree with what @ramonza says. I don't think we should merge this PR.